# feat: optimize getSchedule handler by replacing dayjs with native Date operations