# feat: optimize getSchedule handler by replacing dayjs with native Date operations

## Summary

This PR addresses a critical performance bottleneck in the `getSchedule` tRPC handler where dayjs operations were consuming 4s of CPU time out of 5s total wall time. The optimization replaces dayjs with native JavaScript Date operations across the scheduling pipeline to eliminate micro-task overhead and garbage collection pressure.

**Key Changes:**
- Created `packages/lib/date-utils-native.ts` with optimized native Date utility functions
- Replaced dayjs operations in core scheduling files:
  - `packages/lib/slots.ts`: slot generation and boundary calculations  
  - `packages/lib/date-ranges.ts`: date range processing and working hours
  - `packages/trpc/server/routers/viewer/slots/util.ts`: AvailableSlotsService
- Systematically replaced `.add()`, `.diff()`, `.startOf()`, `.endOf()` with native Date arithmetic
- Used `Date.getTime()` for timestamp comparisons and `Math.floor()` for duration calculations

**Performance Strategy:**
- Replace `dayjs().add(n, 'minutes')` with `new Date(date.getTime() + n * 60 * 1000)`
- Replace `dayjs().diff(other, 'minutes')` with `(date1.getTime() - date2.getTime()) / (1000 * 60)`
- Use direct millisecond arithmetic instead of dayjs timezone-aware operations where safe

## Review & Testing Checklist for Human

**⚠️ HIGH RISK - Critical Business Logic Changes**

- [ ] **Timezone handling correctness**: Verify that timezone conversions still work correctly, especially around DST boundaries and edge cases
- [ ] **End-to-end booking flow testing**: Test actual booking creation through the UI to ensure scheduling still works correctly  
- [ ] **Performance validation**: Measure actual performance improvement in the getSchedule handler with benchmarks
- [ ] **Date arithmetic edge cases**: Test behavior around month boundaries, leap years, and DST transitions
- [ ] **Slot generation accuracy**: Compare slot availability before/after changes to ensure no regression in booking logic

**Recommended Test Plan:**
1. Run full test suite and verify all scheduling tests pass
2. Test booking creation in multiple timezones (especially around DST transitions)
3. Profile getSchedule handler performance before/after to validate improvement
4. Manual testing of availability checking for various date ranges
5. Stress test with high booking volumes to ensure no performance regression

---

### Diagram

```mermaid
%%{ init : { "theme" : "default" }}%%
graph TD
    A["packages/trpc/server/routers/viewer/slots/getSchedule.handler.ts"]:::context --> B["packages/trpc/server/routers/viewer/slots/util.ts<br/>AvailableSlotsService"]:::major-edit
    B --> C["packages/lib/slots.ts<br/>buildSlotsWithDateRanges()"]:::major-edit  
    B --> D["packages/lib/date-ranges.ts<br/>processWorkingHours()"]:::major-edit
    C --> E["packages/lib/date-utils-native.ts<br/>NEW: Native date utilities"]:::major-edit
    D --> E
    B --> F["Various scheduling files<br/>(routing-forms, webhooks, etc.)"]:::minor-edit
    
    subgraph Legend
        L1["Major Edit"]:::major-edit
        L2["Minor Edit"]:::minor-edit  
        L3["Context/No Edit"]:::context
    end

    classDef major-edit fill:#90EE90
    classDef minor-edit fill:#87CEEB
    classDef context fill:#F5F5F5
```

### Notes


- **Link to Devin run**: https://app.devin.ai/sessions/9572e0dd1b1749b98a89cdaeed2ccfda
- **Requested by**: @keithwillcode
- **Performance Context**: Original profiling showed 4s CPU time out of 5s wall time was dayjs-related
- **Compatibility**: Preserved existing function signatures and return types for backward compatibility
- **Timezone Strategy**: Kept dayjs for complex timezone operations where business logic requires it, only replaced performance-critical arithmetic operations

**⚠️ Important**: This PR touches critical scheduling logic. While type checking passes, thorough manual testing is essential before merging to production.